udpated rewards

+++ backend/database/connection/connection.py
@@ -14,11 +14,12 @@ class DatabaseConnection:
 
     async def get_session(self) -> AsyncSession:
         return AsyncSession(bind=self._engine)
-    
+        
     async def __call__(self):
         async with self._engine.begin() as conn:
-            await conn.run_sync(Base.metadata.drop_all)
-            await conn.run_sync(Base.metadata.create_all)
+            if DB_CONFIG.DROP_DB:
+                await conn.run_sync(Base.metadata.drop_all)
+                await conn.run_sync(Base.metadata.create_all)
         await test_db(await self.get_session())
         return self
